separa função e add teste com username
